add card delete and update and delete when parent is deleted

+++ src/controllers/listController.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express'
-import { List } from '../models'
+import { Card, List } from '../models'
 
 export const createList = (req: Request, res: Response): void => {
   List.create(req.body)
@@ -50,9 +50,12 @@ export const updateList = async (req: Request, res: Response): Promise<void> => 
 
 export const deleteList = async (req: Request, res: Response): Promise<void> => {
   try {
+    const list = await List.findById(req.params.id)
+    if (list === null) return
+    list.cards.forEach((card) => {
+      Card.findByIdAndDelete(card).catch((err) => console.log(err))
+    })
     await List.findByIdAndDelete(req.params.id)
-    // TODO delete cards
-    // await Board.deleteMany({ project: req.params.id })
 
     res.status(204).json({
       status: 'Success',
